universe-backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.3
django-filter>=25.1
channels>=4.2
daphne>=4.2
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.2
sentence-transformers>=5.2
faiss-cpu>=1.9
